Refactored PresetMangaerScreen into a seperate PresetList component, added safeareaviews
