+++ package.json
@@ -5,8 +5,7 @@
     "start": "expo start --dev-client",
     "android": "expo run:android",
     "ios": "expo run:ios",
-    "web": "expo start --web",
-    "build:ios": "react-native bundle --entry-file='index.js' --bundle-output='./ios/main.jsbundle' --dev=false --platform='ios'"
+    "web": "expo start --web"
   },
   "dependencies": {
     "@config-plugins/react-native-ble-plx": "^6.0.0",
